+++ src/main/java/duke/command/ByeCommand.java
@@ -1,8 +1,9 @@
-package Duke.Command;
+package duke.command;
 
-import Duke.Storage;
-import Duke.Task.TaskList;
-import Duke.Ui;
+import duke.Storage;
+import duke.Ui;
+
+import duke.task.TaskList;
 
 public class ByeCommand extends Command {
     public ByeCommand() {
