+++ src/main/java/duke/command/DeleteCommand.java
@@ -1,10 +1,12 @@
-package Duke.Command;
+package duke.command;
 
-import Duke.Exception.InvalidArgumentException;
-import Duke.Storage;
-import Duke.Task.Task;
-import Duke.Task.TaskList;
-import Duke.Ui;
+import duke.Ui;
+import duke.Storage;
+
+import duke.task.Task;
+import duke.task.TaskList;
+
+import duke.exception.InvalidArgumentException;
 
 public class DeleteCommand extends Command {
     int positionToDelete;
