+++ src/main/java/duke/exception/InvalidCommandException.java
@@ -1,7 +1,6 @@
-package Duke.Exception;
+package duke.exception;
 
 public class InvalidCommandException extends DukeException {
-
     public InvalidCommandException(String message) {
         super(message);
     }
