+++ src/main/java/duke/task/TaskList.java
@@ -1,9 +1,10 @@
-package Duke.Task;
+package duke.task;
 
 import java.util.ArrayList;
 
 public class TaskList {
     private static ArrayList<Task> taskList;
+
     public TaskList() {
         taskList = new ArrayList<>();
     }
@@ -27,14 +28,17 @@ public class TaskList {
     public Task getTaskAtPosition(int position) {
         return this.taskList.get(position - 1);
     }
+
     public void markTask(int position) {
             Task task = taskList.get(position - 1);
             task.setIsDone(true);
     }
+
     public void unmarkTask(int position) {
         Task task = taskList.get(position - 1);
         task.setIsDone(false);
     }
+
     public void deleteTaskAtPosition(int position) {
         taskList.remove(position - 1);
     }
