+++ src/main/java/duke/task/Todo.java
@@ -1,17 +1,12 @@
-package Duke.Task;
-
-import Duke.Task.Task;
+package duke.task;
 
 public class Todo extends Task {
-
     public Todo(String description) {
         super(description);
     }
-
     public Todo(String description, Boolean isDone) {
         super(description, isDone);
     }
-
     @Override
     public String toString() {
         return "[T]" + this.getStatusIcon() + this.description;
